Change the strategy for managing ssh keys
csync2 doesn't check ssl keys in the usual CA manner. Instead, it
unconditionally accepts a key the first time it sees it, and stores it in
its database. This implies that we need to store the generated key on the
puppet master, otherwise we'll be scuppered if we ever do a rebuild.